Add portions on the CDR loop based on Vgene usage

This function will use the Vgene to add the amino acid sequence of the CDR loop for given clones. For example, recovering the CDR1 and CDR2 sequences, will produce a string with CDR1-CDR2-CDR3 (ex: "SGH.......RS-YFS....ETQ-CASSLTDRTYEQYF"). Sequences are derived from IMGT-based nomenclature. If a vgene is not present or does not match IMGT, unrecoverable regions will be denoted with "!" (ex: "!-!-CASSLTDRTYEQYF").

inferCDR(df, chain = "TRB", regions = c("CDR1", "CDR2"), species = "human")

Arguments

df

The product of combineTCR, combineBCR, or combineExpression.

chain

indicate if both or a specific chain should be used - e.g. "both", "TRA", "TRG", "IGH", "IGL"

regions

CDR loop regions - "FR1", "CDR1", "FR2", "CDR2", and/or "FR3"

species

The species of the experiment - "human" or "mouse"

Value

vector of sequences corresponding with the additional regions

Examples

#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
TRB_CDR123 <- inferCDR(combined[[1]], 
                       chain = "TRB", 
                       regions = c("CDR1", "CDR2"), 
                       species = "human")
#> Error in df[[i]][, cloneCall]: incorrect number of dimensions