Remove TCR and BCR genes from variable gene results

Most single-cell workflows use highly-expressed and highly-variable genes for the initial calculation of PCA and subsequent dimensional reduction. This function will remove the TCR and/or BCR genes from the variable features in a Seurat object or from a vector (potentially generated by the Bioconductor scran workflow).

Usage

quietVDJgenes(input.data, ...)

quietTCRgenes(input.data, ...)

# Default S3 method
quietTCRgenes(input.data, ...)

# S3 method for class 'Seurat'
quietTCRgenes(input.data, assay = NULL, ...)

quietBCRgenes(input.data, ...)

# Default S3 method
quietBCRgenes(input.data, ...)

# S3 method for class 'Seurat'
quietBCRgenes(input.data, assay = NULL, ...)

Arguments

input.data

Single-cell object in Seurat format or vector of variable genes to use in reduction

...

Reserved for future arguments

assay

The Seurat assay slot to use to remove immune receptor genes from, NULL value will default to the default assay

Value

Seurat object or vector list with TCR genes removed.

Author

Nicky de Vrij, Nikolaj Pagh, Nick Borcherding, Qile Yang

Examples

example <- quietVDJgenes(scRep_example)
scRep <- quietTCRgenes(scRep_example)
ibex_example <- quietBCRgenes(scRep_example)