scRepertoire: A toolkit for single-cell immune receptor profiling

scRepertoire is a toolkit for processing and analyzing single-cell T-cell receptor (TCR) and immunoglobulin (Ig). The scRepertoire framework supports use of 10x, AIRR, BD, MiXCR, TRUST4, and WAT3R single-cell formats. The functionality includes basic clonal analyses, repertoire summaries, distance-based clustering and interaction with the popular Seurat and SingleCellExperiment/Bioconductor R single-cell workflows.

See also

Author

Maintainer: Nick Borcherding ncborch@gmail.com

Authors:

Qile Yang qile.yang@berkeley.edu

Ksenia Safina safina@broadinstitute.org

Other contributors:

Justin Reimertz justin.reimertz@childrens.harvard.edu [contributor]