Loading the contigs derived from single-cell sequencing

This function generates a contig list and formats the data to allow for function with combineTCR or combineBCR. If using data derived from filtered outputs of 10X Genomics, there is no need to use this function as the data is already compatible.

loadContigs(input, format = "10X")

Arguments

input

The directory in which contigs are located or a list with contig elements

format

The format of the single-cell contig, currently supporting: "10X", "AIRR", "BD", "JSON", "MiXCR", "Omniscope", "TRUST4", and "WAT3R"

Value

List of contigs for compatibility with combineTCR or combineBCR

Details

The files that this function parses includes:

10X = "filtered_contig_annotation.csv"

AIRR = "airr_rearrangement.tsv"

BD = "Contigs_AIRR.tsv"

Immcantation = "data.tsv"

JSON = ".json"

MiXCR = "clones.tsv"

Omniscope = ".csv"

TRUST4 = "barcode_report.tsv"

WAT3R = "barcode_results.csv"

Examples

TRUST4 <- read.csv("https://www.borch.dev/uploads/contigs/TRUST4_contigs.csv")
contig.list <- loadContigs(TRUST4, format = "TRUST4")

BD <- read.csv("https://www.borch.dev/uploads/contigs/BD_contigs.csv")
contig.list <- loadContigs(BD, format = "BD")

WAT3R <- read.csv("https://www.borch.dev/uploads/contigs/WAT3R_contigs.csv")
contig.list <- loadContigs(WAT3R, format = "WAT3R")